% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonise.R
\name{carbonise}
\alias{carbonise}
\title{carbonise}
\usage{
carbonise(x, habitats)
}
\arguments{
\item{x}{habitats data frame}

\item{habitats}{name of the column containing priority habitats types}
}
\value{
sf features with stored above ground carbon and sequestered carbon per year for each feature
}
\description{
Calculate the carbon storage (above ground biomass and soil carbon) and sequestration for UK priority habitats.
}
\details{
The function calculates the amount of above ground carbon stored in each priority habitat type using data taken from two papers (see main package description for references).
It also calculates stored soil carbon using data from the Global Soil Organic Carbon map.  The soil carbon data only covers the UK.  You need to supply the name of the field with the UK priority habitat descriptions in.
The function can be slow if you have a lot of habitat features, as it summarises soil carbon data from a large raster.
The input dataset needs to be an sf features dataset.
The CRS for this package is EPSG:27700.  You may need to transform your data for the function to work.
}
\examples{
c <- carbonise(x, habitats = "S41Habitat")
}
\seealso{
\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[sf]{geos_measures}}
}
